beer_explorer/requirements.txt
Django
python-dotenv
psycopg2-binary
djangorestframework
django-cors-headers
Pillow
gunicorn
pandas